refactor: UserInfo type 파일로 분리
